Remove TestApplicationStartInfo, add a DefaultApplicationStartInfo
